requirements.txt
supabase
python-dotenv
playwright
beautifulsoup4
requests
